Change joinmarket.cfg to "proxy=false"
Blockr lookups are not using the proxy from joinmarket.cfg and Tails will not let any connection out without tor. So "torify" has to be used for blockr lookup. When using sendpayment.py, or any other script which also joins the irc, the script first does an address lookup (torify needed) and then connects to irc (proxy in .cfg). Double tor does not work. So we'll have to rely on torify solely for now as a workaround until the proxy settings are used for all outside connections.